services/authentification/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy[asyncio]
asyncpg
aiokafka
arq
redis
pyjwt
bcrypt
geoip2
user-agents
aiosmtplib
pydantic
pydantic-settings
pydantic[email]
alembic
prometheus-fastapi-instrumentator
jsonschema
python-multipart
fastapi-limiter
cryptography